% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anc_data_list.R
\name{create_anc_data_list.R}
\alias{create_anc_data_list.R}
\title{Creates an `anc_data` object.}
\usage{
create_anc_data_list.R(anc_prop_by_pos)
}
\description{
Parses relevant data and performs necessary calculations for ancestry-based selection analysis given a tibble of `.anc_prop_by_pos` data.
}
\details{
Currently only works with output from `read_anc_prop_by_pos()`. This function:
* Parses ancestries of interest
* Calculates total ancestry counts
* Calculates global and local ancestry proportions
* Calculates differences in proportion
* Calculates genome-wide mean + sd for diff in proportion
* Calculates mmpp percentile
}
